requirements.txt
psycopg2-binary==2.8.*
environs==9.3.*
discord.py==1.6.*
jikanpy==4.3.*